refactor: moved folders/ delete files
